Refactor to make client_utils and utils.c
